1) EDA and Statistical Analysis of FIFA Game Data to uncover Interesting patterns. 2). Churn prediction using data cleaning, EDA and Statistics for Trend Analysis.
FIFA is a renowned game across the globe and has tons of data, if we use our statistical prowess, we can definitely find interesting patterns. Here, we applied EDA coupled with Statistical Analysis to test, validate and mine multiple hypotheses about game, players and so on. A very popular Telecom Sector case study is Churn Prediction where we want to identify who will churn/won’t. In this case study, we worked on Data Cleaning, EDA along with Statistics to mine interesting patterns and trends.